requirements.txt
crewai
langchain_community
streamlit
streamlit-folium
geopy
fpdf2
pydantic